EVG-15210: Fix cyclical dependency err
